% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{foo}
\alias{foo}
\title{Foo Function to Create a Fancy List}
\usage{
foo(n)
}
\arguments{
\item{n}{number of rows we have in the resulting first element data frame.}
}
\value{
A list with two elements, the first is ... the second is ...
}
\description{
This is just a dummy function.
}
\examples{
foo(2)
}
