% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seoul-package.R
\docType{package}
\name{seoul-package}
\alias{seoul}
\alias{seoul-package}
\title{seoul: Example Package for Seoul Short Course}
\description{
This is a cool package written by Jun Kim.
}
\author{
\strong{Maintainer}: Jun Kim \email{ikarus702@gmail.com}

Authors:
\itemize{
  \item Jun Kim \email{ikarus702@gmail.com}
}

}
\keyword{internal}
